Add glossary page to doc/user-guide * add Documentation Markdown/Main and WithJSX components * add template for creating doc pages with jsx instead of markdown * add glossary page to pages directory
